Remove bulk of full/edit handlig and update test
