+++ src/Features/LanguageServer/ProtocolUnitTests/SemanticTokens/AbstractSemanticTokensTests.cs
@@ -19,20 +19,6 @@ namespace Microsoft.CodeAnalysis.LanguageServer.UnitTests.SemanticTokens
 {
     public abstract class AbstractSemanticTokensTests : AbstractLanguageServerProtocolTests
     {
-        private protected static async Task<LSP.SemanticTokens> RunGetSemanticTokensAsync(TestLspServer testLspServer, LSP.Location caret)
-        {
-            var result = await testLspServer.ExecuteRequestAsync<LSP.SemanticTokensParams, LSP.SemanticTokens>(LSP.Methods.TextDocumentSemanticTokensFullName,
-                CreateSemanticTokensParams(caret), new LSP.VSInternalClientCapabilities(), null, CancellationToken.None);
-            Contract.ThrowIfNull(result);
-            return result;
-        }
-
-        private static LSP.SemanticTokensParams CreateSemanticTokensParams(LSP.Location caret)
-            => new LSP.SemanticTokensParams
-            {
-                TextDocument = new LSP.TextDocumentIdentifier { Uri = caret.Uri }
-            };
-
         private protected static async Task<LSP.SemanticTokens> RunGetSemanticTokensRangeAsync(TestLspServer testLspServer, LSP.Location caret, LSP.Range range)
         {
             var result = await testLspServer.ExecuteRequestAsync<LSP.SemanticTokensRangeParams, LSP.SemanticTokens>(LSP.Methods.TextDocumentSemanticTokensRangeName,
@@ -48,28 +34,9 @@ namespace Microsoft.CodeAnalysis.LanguageServer.UnitTests.SemanticTokens
                 Range = range
             };
 
-        private protected static async Task<SumType<LSP.SemanticTokens, LSP.SemanticTokensDelta>> RunGetSemanticTokensEditsAsync(TestLspServer testLspServer, LSP.Location caret, string previousResultId)
-        {
-            return await testLspServer.ExecuteRequestAsync<LSP.SemanticTokensDeltaParams, SumType<LSP.SemanticTokens, LSP.SemanticTokensDelta>>(LSP.Methods.TextDocumentSemanticTokensFullDeltaName,
-                CreateSemanticTokensParams(caret, previousResultId), new LSP.VSInternalClientCapabilities(), null, CancellationToken.None);
-        }
-
-        private static LSP.SemanticTokensDeltaParams CreateSemanticTokensParams(LSP.Location caret, string previousResultId)
-            => new LSP.SemanticTokensDeltaParams
-            {
-                TextDocument = new LSP.TextDocumentIdentifier { Uri = caret.Uri },
-                PreviousResultId = previousResultId
-            };
-
-        protected static async Task UpdateDocumentTextAsync(string updatedText, Workspace workspace)
-        {
-            var docId = ((TestWorkspace)workspace).Documents.First().Id;
-            await ((TestWorkspace)workspace).ChangeDocumentAsync(docId, SourceText.From(updatedText));
-        }
-
         // VS doesn't currently support multi-line tokens, so we want to verify that we aren't
         // returning any in the tokens array.
-        protected static async Task VerifyNoMultiLineTokens(TestLspServer testLspServer, int[] tokens)
+        private protected static async Task VerifyNoMultiLineTokens(TestLspServer testLspServer, int[] tokens)
         {
             var document = testLspServer.GetCurrentSolution().Projects.First().Documents.First();
             var text = await document.GetTextAsync().ConfigureAwait(false);
